Streamlit_app_aws/requirements.txt
langchain
langchain-community
streamlit
python-dotenv